pyproject.toml
[build-system]
requires = ["setuptools>=61.0.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]

[project]
name = "onomatool"
version = "0.1.0"
description = "AI-powered file renaming tool"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "markitdown",
    "openai",
    "google-generativeai",
    "tomli",
]

[project.scripts]
onomatool = "onomatool.cli:main"

[tool.ruff]
line-length = 88
indent-width = 4
target-version = "py311"

# Minimal exclude patterns
exclude = [
    ".git",
    ".venv",
    "__pycache__",
    "*.pyc",
]

[tool.ruff.lint]
# Basic rules plus import sorting and some useful quality checks
select = ["E", "F", "W", "I", "B", "UP", "C4"]
ignore = [
    "E501",   # line-too-long (let formatter handle this)
]

[tool.ruff.lint.isort]
known-first-party = ["onomatool"]
combine-as-imports = true
split-on-trailing-comma = true

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
line-ending = "lf"
